[BY-267] Move: delete folders inside authenticate
