+++ lib/constants/constants.dart
@@ -3,7 +3,7 @@ constants.dart : 앱에서 사용하는 각종 상수 값을 정리한 파일
 - 상수명 lowerCamelCase 스타일 따름 (Dart 권장 : PREFER using lowerCamelCase for constant names)
 */
 
-import 'package:byourside/screen/authenticate/policy/policy.dart';
+import 'package:byourside/screen/authenticate/policy.dart';
 import 'package:byourside/screen/mypage/my_block_user.dart';
 import 'package:byourside/screen/mypage/my_community_post.dart';
 import 'package:byourside/screen/mypage/my_scrap.dart';
